wu-move-from-ipc: Move update_weights_from_ipc onto WeightUpdater
Refactor chain ID: wu-move-from-ipc
