+++ python/sglang/srt/managers/tp_worker.py
@@ -172,7 +172,9 @@ class BaseTpWorker(ABC):
 
     def update_weights_from_ipc(self, recv_req: UpdateWeightsFromIPCReqInput):
         """Update weights from IPC for checkpoint-engine integration."""
-        success, message = self.model_runner.update_weights_from_ipc(recv_req)
+        success, message = self.model_runner.weight_updater.update_weights_from_ipc(
+            recv_req
+        )
         return success, message
 
     def get_weights_by_name(self, recv_req: GetWeightsByNameReqInput):
