+++ python/sglang/srt/model_executor/model_runner.py
@@ -3113,23 +3113,6 @@ class ModelRunner(ModelRunnerKVCacheMixin):
     def check_weights(self, action: str):
         return self._weight_checker.handle(action=action)
 
-    def update_weights_from_ipc(self, recv_req):
-        """Update weights from IPC for checkpoint-engine integration."""
-        try:
-            from sglang.srt.checkpoint_engine.checkpoint_engine_worker import (
-                SGLangCheckpointEngineWorkerExtensionImpl,
-            )
-
-            # Create a worker extension that integrates with SGLang's model
-            worker = SGLangCheckpointEngineWorkerExtensionImpl(self)
-            worker.update_weights_from_ipc(recv_req.zmq_handles)
-            return True, "IPC weight update completed successfully"
-        except ImportError as e:
-            return False, f"IPC weight update failed: ImportError {e}"
-        except Exception as e:
-            logger.error(f"IPC weight update failed: {e}")
-            return False, str(e)
-
     def _maybe_rebalance_after_rank_fault(
         self,
         output: ModelRunnerOutput,
